+++ lib/arkit_plugin.dart
@@ -1,3 +1,4 @@
 export 'package:arkit_plugin/widget/arkit_scene_view.dart';
+export 'package:arkit_plugin/widget/arkit_arplane_detection.dart';
 export 'package:arkit_plugin/geometries/arkit_position.dart';
 export 'package:arkit_plugin/geometries/arkit_sphere.dart';