+++ lib/widget/arkit_arplane_detection.dart
@@ -1,0 +1,11 @@
+/// Indicating the type of planes to detect.
+enum ARPlaneDetection {
+  /// No plane detection is run
+  none,
+
+  /// Plane detection determines horizontal planes in the scene.
+  horizontal,
+
+  /// Plane detection determines vertical planes in the scene. (iOS >= 11.3)
+  vertical,
+}