+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog
+
+## 0.0.2
+
+* Plane detection enum implemented
+* ARKitSceneView.autoenablesDefaultLighting property implemented
+* Documentation added
+
 ## 0.0.1
 
 * PoC of the ARKit port